cleanup disassembler utility code

+++ bld/ndisasm/stand/c/formasm.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2019 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2022 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -96,7 +96,7 @@ bool IsMasmOutput( void )
 
 static void printRawAndAddress( const char *raw_data, dis_sec_offset address )
 {
-    int     len;
+    size_t  len;
 
     len = strlen( raw_data );
     switch( len ) {
